@@ -1,10 +1,19 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.0.7'
+gem 'rails', '3.1.0'
 gem 'paperclip', '2.3.11'
 gem 'rmagick'
+# gem 'jquery-rails'
 
 gem 'sqlite3'
+
+# Gems used only for assets and not required
+# in production environments by default.
+group :assets do
+  gem 'sass-rails', "  ~> 3.1.0"
+  gem 'coffee-rails', "~> 3.1.0"
+  gem 'uglifier'
+end
 
 group :development, :test do
   gem 'rspec-rails'
@@ -12,8 +21,10 @@ group :development, :test do
 end
 
 group :test do
+  gem 'jasmine'
   gem 'capybara', '1.0.0'
   gem 'launchy'
   gem 'spork', '~> 0.9.0.rc2'
   gem 'database_cleaner'
 end
+
