@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-AngularDemo::Application.config.session_store :cookie_store, :key => '_angular_demo_session'
+AngularDemo::Application.config.session_store :cookie_store, key: '_angular_demo_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
 # (create the session table with "rails generate session_migration")
-AngularDemo::Application.config.session_store :active_record_store
+# AngularDemo::Application.config.session_store :active_record_store
