@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-AngularDemo::Application.config.secret_token = '63632ed497963cfab03b02aec47cc2855344e296eba32a5074ddf75a29bcebad1a87b3a481bb60627e160989129edba8c394b37ecf63425e88abcb3f34c2845b'
+AngularDemo::Application.config.secret_token = '242a2670e67f8f1bbe39cd1312db282aa144c470c21d9813eef907ffcde3bd486a033e551927bd530c97bbafdc5756e0c292b53ee80814407d5430812f68b268'
